fix(agent-testing): allow user to stop agent when typing or schema are not set

+++ packages/agent-nodejs-testing/src/integrations/testable-agent.ts
@@ -35,12 +35,14 @@ export default class TestableAgent<TypingsSchema extends TSchema = TSchema> exte
 
   async stop(): Promise<void> {
     await this.agent.stop();
-    await fs.rm(this.agentOptions.typingsPath, { force: true });
-    await fs.rm(this.agentOptions.schemaPath, { force: true });
+    if (this.agentOptions.typingsPath) await fs.rm(this.agentOptions.typingsPath, { force: true });
+    if (this.agentOptions.schemaPath) await fs.rm(this.agentOptions.schemaPath, { force: true });
   }
 
   async start(): Promise<void> {
     await this.agent.mountOnStandaloneServer(this.port).start();
+    if (!this.agentOptions.schemaPath) throw new Error('schemaPath is required');
+
     this.schema = JSON.parse(await fs.readFile(this.agentOptions.schemaPath, 'utf8'));
   }
